+++ npm-packages/shared-utils/src/types/types.ts
@@ -50,6 +50,11 @@ export enum IMCMessageTypeEnum {
   EditorOpenLink = "editor-open-link",
   // Open app in editor
   EditorOpenApp = "editor-open-app",
+  // Locale
+  EditorLocaleUpdate = "editor-locale-update",
+  EditorAppRequestLocale = "editor-app-request-locale",
+  // Artifact update
+  EditorArtifactUpdate = "editor-artifact-update",
   // #endregion
 
   // #region Platform API interaction messages (require OS-like environment)
@@ -320,3 +325,15 @@ export type ListPathOptions = {
 // input:
 // output:
 // }
+
+export type Artifact = {
+  type: "app";
+  // | "image" | "text" | "audio" | "video";
+  data: AppArtifact;
+};
+
+export type AppArtifact = {
+  appId: string;
+  version: string;
+  sourceUrl?: string;
+};