+++ npm-packages/react-api/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @pulse-editor/react-api
 
+## 0.1.1-beta.72
+
+### Patch Changes
+
+- Add artifact output hook
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-beta.72
+
+## 0.1.1-beta.71
+
+### Patch Changes
+
+- Add locale hook
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-beta.71
+
 ## 0.1.1-beta.70
 
 ### Patch Changes